added authorization middleware

+++ handlers/taskHandler/taskHandler.go
@@ -11,7 +11,7 @@ import (
 	"time"
 )
 
-const contextKeyUserID = "userId"
+const contextKeyId = "id"
 
 type TaskHandler struct {
 	taskService *taskService.TaskService
@@ -27,8 +27,7 @@ func (h *TaskHandler) CreateTask(writer http.ResponseWriter, request *http.Reque
 		utilities.ErrorJsonRespond(writer, http.StatusBadRequest, fmt.Errorf("json decode failed"))
 		return
 	}
-	//TODO: достать id пользователя из контекста запроса (его туда должна положить мидлварь авторизации)
-	userId := "some hex uuid string" // request.Context().Value(contextKeyId).(string)
+	userId := request.Context().Value(contextKeyId).(string)
 
 	task, err := h.taskService.CreateTask(task, userId)
 	if err != nil {
@@ -40,6 +39,11 @@ func (h *TaskHandler) CreateTask(writer http.ResponseWriter, request *http.Reque
 
 func (h *TaskHandler) GetTasksByUserId(writer http.ResponseWriter, request *http.Request) {
 	userId, _ := mux.Vars(request)["id"]
+	contextUserID := request.Context().Value(contextKeyId).(string)
+	if userId != contextUserID {
+		utilities.ErrorJsonRespond(writer, http.StatusForbidden, fmt.Errorf("no acces for this id"))
+		return
+	}
 	var tasks, err = h.taskService.GetByUserId(userId)
 	if err != nil {
 		utilities.ErrorJsonRespond(writer, http.StatusInternalServerError, err)
@@ -50,6 +54,11 @@ func (h *TaskHandler) GetTasksByUserId(writer http.ResponseWriter, request *http
 
 func (h *TaskHandler) DeleteByUserId(writer http.ResponseWriter, request *http.Request) {
 	userId, _ := mux.Vars(request)["id"]
+	contextUserID := request.Context().Value(contextKeyId).(string)
+	if userId != contextUserID {
+		utilities.ErrorJsonRespond(writer, http.StatusForbidden, fmt.Errorf("no acces for this id"))
+		return
+	}
 	if err := h.taskService.DeleteById(userId); err != nil {
 		utilities.ErrorJsonRespond(writer, http.StatusInternalServerError, err)
 		return
@@ -58,8 +67,7 @@ func (h *TaskHandler) DeleteByUserId(writer http.ResponseWriter, request *http.R
 }
 
 func (h *TaskHandler) GetUserTasks(writer http.ResponseWriter, request *http.Request) {
-	//TODO: достать id пользователя из контекста запроса (его туда должна положить мидлварь авторизации)
-	userId := "some hex uuid string" // request.Context().Value(contextKeyId).(string)
+	userId := request.Context().Value(contextKeyId).(string)
 	dateFrom := request.FormValue("dateFrom")
 	dateTo := request.FormValue("dateTo")
 	label := request.FormValue("label")
